refactor signature.py and signature_json.py into signature submodule
